+++ SwiftUsd.docc/AboutThisRepo/ChangeLog.md
@@ -15,6 +15,7 @@ Changes to SwiftUsd
 Released TBD, based on OpenUSD TBD
 - Add `Overlay.HgiGLWrapper`, exposing HgiGL alongside HgiMetal in the Swift bindings
 - Add `Overlay.ExecUsdSystem`, exposing ExecUsdSystem in the Swift bindings
+- Add `Overlay.ArAssetWrapper`, `Overlay.ArWritableAssetWrapper`
 
 ### 7.0.1
 Released 2026-06-04, based on OpenUSD v26.05